+++ es/utils/aepp-wallet-communication/wallet-detector.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-undef */
 /*
  * ISC License (ISC)
  * Copyright (c) 2018 aeternity developers
@@ -105,7 +104,7 @@ function getWallets () {
 export const WalletDetector = AsyncInit.compose({
   async init ({ connection } = {}) {
     if (!window) throw new Error('Window object not found, you can run wallet detector only in browser')
-    this.connection = connection || BrowserWindowMessageConnection({ connectionInfo: { id: 'spy' }})
+    this.connection = connection || BrowserWindowMessageConnection({ connectionInfo: { id: 'spy' } })
   },
   methods: { scan, stopScan, getWallets }
 })