+++ examples/browser/extension/src/js/background.js
@@ -1,12 +1,11 @@
+/* global chrome */
+
 import '../img/icon-128.png'
 import '../img/icon-34.png'
 
-import MemoryAccount from '@aeternity/aepp-sdk/es/account/memory'
-import Node from '@aeternity/aepp-sdk/es/node'
-import { RpcWallet } from '@aeternity/aepp-sdk/es/ae/wallet'
+import { RpcWallet, Node, MemoryAccount, Crypto } from 'AE_SDK_MODULES'
 import BrowserRuntimeConnection
-  from '@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/connection/browser-runtime'
-import { generateKeyPair } from '@aeternity/aepp-sdk/es/utils/crypto'
+  from 'AE_SDK_MODULES/utils/aepp-wallet-communication/connection/browser-runtime'
 
 // const account = MemoryAccount({
 //   keypair: {
@@ -15,7 +14,7 @@ import { generateKeyPair } from '@aeternity/aepp-sdk/es/utils/crypto'
 //   }
 // })
 
-const account2 = MemoryAccount({ keypair: generateKeyPair() })
+const account2 = MemoryAccount({ keypair: Crypto.generateKeyPair() })
 
 // Init accounts
 const accounts = [