+++ examples/browser/extension/utils/webserver.js
@@ -1,35 +1,35 @@
-var WebpackDevServer = require("webpack-dev-server"),
-    webpack = require("webpack"),
-    config = require("../webpack.config"),
-    env = require("./env"),
-    path = require("path");
+var WebpackDevServer = require('webpack-dev-server')
+var webpack = require('webpack')
+var config = require('../webpack.config')
+var env = require('./env')
+var path = require('path')
 
-var options = (config.chromeExtensionBoilerplate || {});
-var excludeEntriesToHotReload = (options.notHotReload || []);
+var options = (config.chromeExtensionBoilerplate || {})
+var excludeEntriesToHotReload = (options.notHotReload || [])
 
 for (var entryName in config.entry) {
   if (excludeEntriesToHotReload.indexOf(entryName) === -1) {
     config.entry[entryName] =
       [
-        ("webpack-dev-server/client?http://localhost:" + env.PORT),
-        "webpack/hot/dev-server"
-      ].concat(config.entry[entryName]);
+        ('webpack-dev-server/client?http://localhost:' + env.PORT),
+        'webpack/hot/dev-server'
+      ].concat(config.entry[entryName])
   }
 }
 
 config.plugins =
-  [new webpack.HotModuleReplacementPlugin()].concat(config.plugins || []);
+  [new webpack.HotModuleReplacementPlugin()].concat(config.plugins || [])
 
-delete config.chromeExtensionBoilerplate;
+delete config.chromeExtensionBoilerplate
 
-var compiler = webpack(config);
+var compiler = webpack(config)
 
 var server =
   new WebpackDevServer(compiler, {
     hot: true,
-    contentBase: path.join(__dirname, "../build"),
-    headers: { "Access-Control-Allow-Origin": "*" },
+    contentBase: path.join(__dirname, '../dist'),
+    headers: { 'Access-Control-Allow-Origin': '*' },
     disableHostCheck: true
-  });
+  })
 
-server.listen(env.PORT);
+server.listen(env.PORT)