+++ examples/browser/vuejs/connect-two-ae/aepp/webpack.config.js
@@ -6,7 +6,7 @@ const { CleanWebpackPlugin } = require('clean-webpack-plugin')
 const { VueLoaderPlugin } = require('vue-loader')
 // const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin
 const PurgecssPlugin = require('purgecss-webpack-plugin')
-let glob = require('glob-all')
+const glob = require('glob-all')
 
 const distFolder = path.resolve(__dirname, 'dist')
 const jsLoader = 'babel-loader!standard-loader?error=true'
@@ -25,14 +25,11 @@ module.exports = {
   mode: process.env.NODE_ENV === 'prod' ? 'production' : 'development',
   resolve: {
     alias: {
-      // use this if you are installing the SDK as dependency
-      // AE_SDK_MODULES: path.resolve(__dirname, 'node_modules/@aeternity/aepp-sdk/es/')
-      // use this, if you are running this app from inside the Aepp-SDK repo/folder
       AE_SDK_MODULES: '../../../../../../../es/'
     }
   },
   entry: {
-    'wallet': './src/index.js'
+    wallet: './src/index.js'
   },
   output: {
     filename: 'bundle.js?[hash]'