+++ examples/browser/vuejs/connect-two-ae/identity/src/components/Home.vue
@@ -55,6 +55,11 @@
         class="w-32 rounded rounded-full bg-purple text-white py-2 px-4 pin-r mr-8 mt-4 text-xs"
         @click="switchNode"
       >Switch Node</button>
+      <button
+        v-if="client"
+        class="w-32 rounded rounded-full bg-purple text-white py-2 px-4 pin-r mr-8 mt-4 text-xs"
+        @click="disconnect"
+      >Disconnect</button>
     </div>
 
     <div v-if="!aeppUrl" class="w-full p-4 h-64 border border-black border-dashed shadow mx-auto mt-4 bg-grey-lighter">
@@ -68,11 +73,10 @@
 </template>
 
 <script>
-  // AE_SDK_MODULES is a webpack alias present in webpack.config.js
-  import { MemoryAccount, RpcWallet, Node } from '@aeternity/aepp-sdk/es'
+  import { MemoryAccount, RpcWallet, Node, Crypto } from 'AE_SDK_MODULES'
   import BrowserWindowMessageConnection
-    from '@aeternity/aepp-sdk/es/utils/aepp-wallet-communication/connection/browser-window-message'
-  import { generateKeyPair } from '@aeternity/aepp-sdk/es/utils/crypto'
+    from 'AE_SDK_MODULES/utils/aepp-wallet-communication/connection/browser-window-message'
+  import { METHODS } from 'AE_SDK_MODULES/utils/aepp-wallet-communication/schema'
 
   const errorAsField = async fn => {
     try {
@@ -123,6 +127,10 @@
       disconnect () {
         const { clients: aepps } = this.client.getClients()
         const aepp = Array.from(aepps.values())[0]
+        aepp.sendMessage({
+          method: METHODS.closeConnection,
+          params: { reason: 'bye' },
+        }, true)
         aepp.disconnect()
       },
       async switchAccount () {
@@ -138,7 +146,7 @@
       }
     },
     async created () {
-      const account2 = MemoryAccount({ keypair: generateKeyPair() })
+      const account2 = MemoryAccount({ keypair: Crypto.generateKeyPair() })
       const testNetNode = await Node({ url: this.url })
       const mainNetNode = await Node({ url: this.mainNetUrl })
 
@@ -146,8 +154,8 @@
         if (confirm(`Client ${aepp.info.name} with id ${aepp.id} want to ${getActionName(params)}`)) accept()
         else deny()
       }
-      const keypair = generateKeyPair()
-      const keypair2 = generateKeyPair()
+      const keypair = Crypto.generateKeyPair()
+      const keypair2 = Crypto.generateKeyPair()
       const sdkAcc = this.publicKey
       this.client = await RpcWallet({
         nodes: [